PowerShell scripts for parse from file/add/remove Twitter accounts.
PowerShell script to build an input file out of CSV with Twitter
handles. Splits the handles in batches of 100 which is the API limit and
ignores any invalid (non-alphanumeric) handles.
PowerShell script to read batches of handles from an input file and
upload them to the list.
Providing the array of user handles and the action type as parameters.
Ability to remove an array of user handles from a twitter list.